% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSpatialCommunStrength.R
\name{AdjustCommunStrength}
\alias{AdjustCommunStrength}
\title{Adjust Communication Strength Based on Spatial Distance}
\usage{
AdjustCommunStrength(prob, coord, neighbors, n.distance)
}
\arguments{
\item{prob}{A 3D array representing communication probabilities between cells.}

\item{coord}{A matrix of spatial coordinates for each cell. Rows are cell names, and columns are spatial dimensions.}

\item{neighbors}{A named list where each element contains the nearest neighbors of a cell, as determined by a k-nearest neighbors algorithm.}

\item{n.distance}{Numeric, the spatial distance constant for adjusting communication strength.}
}
\value{
A 3D array representing communication probabilities between cells, adjusted based on spatial distance.
}
\description{
This function adjusts the communication strength between cells based on their spatial distance.
}
