% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAvgExp.R
\name{CalculateAvgExp}
\alias{CalculateAvgExp}
\title{Calculate the average expression.}
\usage{
CalculateAvgExp(
  pcc_obj,
  input.data = NULL,
  methods = NULL,
  return.object = TRUE
)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{input.data}{A sparse matrix, using \code{data.signaling} slot as default.}

\item{methods}{The method(average,median,quartile) used to calculate the average expression.}

\item{return.object}{Logical, whether to return the updated PlantCellChat object.}
}
\value{
The updated PlantCellChat object with the average expression in the 'data.average' slot.
}
\description{
Providing a pipeline to calculate the average gene expression in
different cells,including three methods(average,median,quartile).The quartile
method using formula "1/2Q2+1/4(Q1+Q3)" to calculate the average expression.
}
\examples{
pcc_obj <- CalculateAvgExp(pcc_obj,methods = "average")
}
