% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{CalculateCommunStrength}
\alias{CalculateCommunStrength}
\title{Calculate Communication Strength}
\usage{
CalculateCommunStrength(
  pcc_obj,
  Kh = 0.5,
  n = 1,
  exp.methods = "average",
  num.permutations = 100,
  seed = 123
)
}
\arguments{
\item{pcc_obj}{A \code{PlantCellChat} object which contains the average expression data (stored in the \code{data.average} slot)
and over-expressed ligand-receptor interaction data (stored in the \code{lrpairs} slot).}

\item{Kh}{Numeric, the dissociation constant in the Hill equation, controlling the sensitivity of ligand-receptor binding.}

\item{n}{Numeric, the Hill coefficient describing the cooperativity of ligand-receptor binding.}

\item{exp.methods}{Character string, specifying the method for calculating average expression, default is "average".}

\item{num.permutations}{Integer, the number of permutations to run in calculating the p-value for communication strength.}

\item{seed}{Integer, the seed for random number generation, ensuring reproducibility.}
}
\value{
The \code{PlantCellChat} object with updated communication strength data stored in the \code{net} slot, including:
- \code{prob}: A 3-dimensional matrix containing calculated communication probabilities for each ligand-receptor pair across clusters.
- \code{pvalue}: A 3-dimensional matrix with p-values based on permutation tests for assessing the significance of each probability.
- \code{df}: A data frame summarizing the calculated probabilities and p-values for each ligand-receptor interaction.
}
\description{
This function using the Law of mass to calculate the communication strength based on
ligands and receptors.
}
