% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{CalculateCommunStrength.Sub}
\alias{CalculateCommunStrength.Sub}
\title{sub-function of \code{CalculateCommunStrength}}
\usage{
CalculateCommunStrength.Sub(pcc_obj, Kh = 0.5, n = 1)
}
\arguments{
\item{pcc_obj}{A \code{PlantCellChat} object which contains the average expression data (stored in the \code{data.average} slot)
and over-expressed ligand-receptor interaction data (stored in the \code{lrpairs} slot).}

\item{Kh}{Numeric, the dissociation constant in the Hill equation, controlling the sensitivity of ligand-receptor binding.}

\item{n}{Numeric, the Hill coefficient describing the cooperativity of ligand-receptor binding.}
}
\value{
The \code{PlantCellChat} object with an updated \code{prob} matrix, stored in the \code{net} slot, which contains the calculated
communication strength probabilities for each ligand-receptor pair across cell clusters.
}
\description{
sub-function of \code{CalculateCommunStrength}
}
