% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{CalculateSignalingStrength}
\alias{CalculateSignalingStrength}
\title{Calculate Signaling Strength}
\usage{
CalculateSignalingStrength(
  pcc_obj,
  Kh = 0.5,
  n = 1,
  exp.methods = "average",
  num.permutations = 100,
  seed = 123
)
}
\arguments{
\item{pcc_obj}{A \code{PlantCellChat} object containing data on average expression (in the \code{data.average} slot) and
over-expressed signaling interactions (in the \code{lrpairs} slot).}

\item{Kh}{Numeric, the dissociation constant in the Hill equation, which modulates the sensitivity of signaling molecule binding.}

\item{n}{Numeric, the Hill coefficient, which describes the cooperativity of signaling molecule binding.}

\item{exp.methods}{Character, the method for calculating average expression; default is "average".}

\item{num.permutations}{Integer, the number of permutations used in p-value calculation, providing statistical significance of the signaling strength.}

\item{seed}{Integer, a seed for random number generation, ensuring reproducibility.}
}
\value{
The \code{PlantCellChat} object with updated signaling strength data stored in the \code{netSignal} slot, including:
- \code{prob}: A 3-dimensional matrix containing calculated signaling probabilities.
- \code{pvalue}: A 3-dimensional matrix with p-values based on permutation tests.
- \code{df}: A data frame summarizing the calculated probabilities and p-values for each signaling interaction.
}
\description{
This function calculates the signaling strength between cell populations based on signaling molecule interactions.
It uses the Law of mass to model cooperative binding effects and performs permutation testing to assess the statistical
significance of each signaling strength value. The function outputs probability and p-value matrices for signaling interactions.
}
