% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{CalculateSignalingStrength.Sub}
\alias{CalculateSignalingStrength.Sub}
\title{sub-function of \code{CalculateSignalingStrength}}
\usage{
CalculateSignalingStrength.Sub(pcc_obj, Kh = 0.5, n = 1, Ls = 1)
}
\arguments{
\item{pcc_obj}{A \code{PlantCellChat} object containing data on average expression (in the \code{data.average} slot) and
over-expressed signaling interactions (in the \code{lrpairs} slot).}

\item{Kh}{Numeric, the dissociation constant in the Hill equation, which modulates the sensitivity of signaling molecule binding.}

\item{n}{Numeric, the Hill coefficient, which describes the cooperativity of signaling molecule binding.}
}
\value{
Returns the modified \code{PlantCellChat} object with an updated \code{prob} matrix stored in the \code{net} slot. This matrix
contains the calculated communication strength probabilities for each ligand-receptor pair across cell clusters.
}
\description{
sub-function of \code{CalculateSignalingStrength}
}
