% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSpatialCommunStrength.R
\name{CalculateSpatialCommunStrength}
\alias{CalculateSpatialCommunStrength}
\title{Calculate Spatial Communication Strength}
\usage{
CalculateSpatialCommunStrength(
  pcc_obj,
  Kh = 0.5,
  n = 1,
  coord,
  k.neighbors = 5,
  n.distance = 10
)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object}

\item{Kh}{Numeric, the dissociation constant in the Hill equation, which modulates the sensitivity of signaling molecule binding.}

\item{n}{Numeric, the Hill coefficient, which describes the cooperativity of signaling molecule binding.}

\item{k.neighbors}{Integer, the number of nearest neighbors to consider in spatial calculations. Default is 5.}

\item{n.distance}{Numeric, the spatial distance constant for adjusting communication strength. Default is 10.}
}
\value{
A PlantCellChat object with updated spatial communication strength information stored in the \code{sp.net} slot.
}
\description{
This function calculates the spatial communication strength between cells based on ligand-receptor interactions.
}
