% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSignalCommunStrength.R
\name{CompareSignalCommunStrength}
\alias{CompareSignalCommunStrength}
\title{Compare Signal Communication Strength}
\usage{
CompareSignalCommunStrength(
  pcc_obj,
  pcc_obj_list = NULL,
  ligand.type = "lrpairs",
  key.signal = NULL,
  key.source = NULL,
  key.target = NULL,
  top.n = 10,
  label.cex = 10,
  title.cex = 11,
  expand.limits = 0,
  input.color = NULL,
  show.title = T,
  coord.flip = F,
  scale.factor = 1
)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{pcc_obj_list}{A list of PlantCellChat objects for comparison across multiple groups}

\item{ligand.type}{A character string specifying the type of ligand, either "lrpairs" (ligand-receptor pairs) or "signal".}

\item{key.signal}{A character string specifying the key signaling molecule to filter the network.}

\item{key.source}{A character string specifying the cell type or tissue where the ligand located.}

\item{key.target}{A character string specifying the cell type or tissue where the receptor located.}

\item{label.cex}{Numeric, scaling factor for vertex label font size.}

\item{title.cex}{Numeric, scaling factor for title font size.}

\item{expand.limits}{Numeric, Expanding the plot limits on the y-axis. Default is 0.}

\item{input.color}{Character. Hex color code for the bars in the plot.}

\item{show.title}{Logical, indicating whether to display the plot title.}
}
\value{
A ggplot object representing the bar plot.
}
\description{
This function compares the communication strength of different groups based on the ligand-receptor pairs or signaling molecules.
}
