% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{correct}
\alias{correct}
\title{Correct Matrix}
\usage{
correct(matrix)
}
\arguments{
\item{matrix}{A matrix to be corrected.}
}
\value{
A corrected matrix.
}
\description{
This function corrects a matrix by replacing NA values
with 1 and setting non-1 values to NA.
}
\examples{
set.seed(123)
matrix <- matrix(runif(11 * 5), nrow = 11, ncol = 5)
matrix[5:7,] <- NA
corrected_matrix <- correct(matrix)
}
