% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlantCellChat.R
\name{CreatePlantCellChat}
\alias{CreatePlantCellChat}
\title{Create PlantCellChat object to analysis cell communication.}
\usage{
CreatePlantCellChat(
  object,
  meta = NULL,
  input.ident = NULL,
  assay = NULL,
  spatial.coord = NULL
)
}
\arguments{
\item{object}{A normalized data matrix or Seurat object.}

\item{meta}{Cell identity information.}

\item{input.ident}{A character name of the variable in meta data, defining cell groups.}

\item{assay}{Assay to use when the input is a Seurat object.}
}
\value{
A Plantcellchat object
}
\description{
This function creates a PlantCellChat object for cell communication analysis,
modified from the package \code{CellChat} https://github.com/sqjin/CellChat/blob/master/R/CellChat_class.R
}
