% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlantCellChat.R
\name{.error_if_no_Seurat}
\alias{.error_if_no_Seurat}
\title{Check if Seurat package is installed.}
\usage{
.error_if_no_Seurat()
}
\value{
The result of judgement.
}
\description{
Check if Seurat package is installed.
}
\examples{
error_if_no_Seurat()
}
