% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{ExtractExpMatrix}
\alias{ExtractExpMatrix}
\title{Extract Expression Matrices}
\usage{
ExtractExpMatrix(
  pcc_obj,
  comps,
  interaction.type,
  na.replace = F,
  zero.replace = F,
  spatial = F
)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{comps}{The component (or column) of the interaction data to extract expression for.}

\item{interaction.type}{A string defining the type of interaction. Accepts either "lri" for
ligand-receptor interactions or "sri" for signaling molecule interactions.}

\item{na.replace}{Logical, if TRUE, replace NA values in the matrix with 1. Default is FALSE.}

\item{zero.replace}{Logical, if TRUE, replace NA values in the matrix with 0. Default is FALSE.}

\item{spatial}{Logical, if TRUE, extract expression data from the \code{data.signaling} slot. Default is FALSE.}
}
\value{
A matrix of expression values for the specified components, with optional replacements.
}
\description{
This function extracts the expression matrix for specific components from
the PlantCellChat object. It allows replacement of NA values or zero values with a specified value.
}
