% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractOverExpressedInteractions.R
\name{ExtractOverExpressedInteractions}
\alias{ExtractOverExpressedInteractions}
\title{Extract over-expressed interactions}
\usage{
ExtractOverExpressedInteractions(pcc_obj)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object which contains the result of differential expression analysis(saved in the 'diffexp' slot).}
}
\value{
The over-expressed ligand-receptor interactions which saved in
the \code{LR} slot of the PlantCellChat object.
}
\description{
Extract over-expressed ligand-receptor interactions from PlantCellChatDB.
}
