% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSignalingData.R
\name{ExtractSignalingData}
\alias{ExtractSignalingData}
\title{Extract subset sparse matrix of ligands and receptors gene which mediating signaling.}
\usage{
ExtractSignalingData(pcc_obj)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}
}
\value{
A sparse matrix recording ligands and receptors for signaling which saved in
the \code{data.signaling} slot of the PlantCellChat object.
}
\description{
Extract subset sparse matrix of ligands and receptors gene which mediating signaling.
}
