% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCommunStrength.R
\name{MergeExpMatrix}
\alias{MergeExpMatrix}
\title{Merge Expression Matrices}
\usage{
MergeExpMatrix(matrix1 = NULL, matrix2 = NULL, matrix_list = NULL)
}
\arguments{
\item{matrix1}{The first expression matrix.}

\item{matrix2}{The second expression matrix.}

\item{matrix_list}{A list of expression matrices to be merged.}
}
\value{
A merged expression matrix.
}
\description{
This function merges expression matrices by filling
missing values with non-missing values from other matrices.
}
\examples{
matrix1 <- matrix(c(NA,2,4,NA,7,9,NA,4,11),nrow = 3)
matrix2 <- matrix(c(5,NA,6,21,NA,9,8,NA,5),nrow = 3)
matrix3 <- matrix(c(1,3,NA,7,4,NA,6,11,NA),nrow = 3)

merged_matrix <- MergeExpMatrix(matrix1,matrix2)
merged_matrix <- MergeExpMatrix(matrix_list = list(matrix1,matrix2,matrix3))

}
\seealso{
\code{\link[stats]{complete.cases}}
}
