% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PccIdents.R
\name{PccIdents}
\alias{PccIdents}
\title{Create a shortcut command to print PlantCellChat ident.}
\usage{
PccIdents(pcc_obj)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}
}
\value{
The PlantCellChat defalut ident.
}
\description{
Create a shortcut command to print PlantCellChat ident.
}
\examples{
PccIdents(pcc_obj)
}
