% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlantCellChat.R
\docType{class}
\name{PlantCellChat-class}
\alias{PlantCellChat-class}
\alias{PlantCellChat}
\title{Create PlantCellChat class}
\description{
Create PlantCellChat class
}
\section{Slots}{

\describe{
\item{\code{data.raw}}{raw count data matrix}

\item{\code{data}}{normalized data matrix for PlantCellChat analysis (Genes should be in rows and cells in columns)}

\item{\code{data.average}}{A matrix of average expression of each cell group}

\item{\code{data.signaling}}{a subset of normalized matrix only containing signaling genes}

\item{\code{net}}{Three-dimensional matrix (K x K x N) representing the communication strength between different cell group for each ligand-receptor pairs}

\item{\code{netSignal}}{Three-dimensional(K x K x M) array recording the communication strength between different cell group for signals.}

\item{\code{database}}{ligand-receptor interaction database used in the analysis}

\item{\code{lrpairs}}{subset data of ligand-receptor interaction database for signaling}

\item{\code{meta}}{data frame storing the information associated with each cell}

\item{\code{idents}}{the default label defining the cell identity used for all analysis}

\item{\code{diffexp}}{A list: one element is a vector consisting of the identified over-expressed signaling genes; one element is a data frame returned from the differential expression analysis}

\item{\code{spatial}}{description A list containing spatial information for cells, such as spatial coordinates.}
}}

