% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingCommunNetwork.R
\name{PlottingCommunNetwork}
\alias{PlottingCommunNetwork}
\title{Plot Cell Communication Network}
\usage{
PlottingCommunNetwork(
  pcc_obj,
  ligand.type,
  comm.pattern = NULL,
  key.lrpairs = NULL,
  key.signal = NULL,
  key.cluster = NULL,
  edge.size = 8,
  edge.alpha = 0.5,
  arrow.size = 0.2,
  arrow.width = 0.6,
  label.cex = 1,
  label.distance = -0.5,
  title.cex = 1.2,
  show.title = T,
  input.color = NULL,
  plot.type = "circle",
  normalize = T
)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{ligand.type}{A character string specifying the type of ligand, either "lrpairs" (ligand-receptor pairs) or "signal".}

\item{comm.pattern}{A character string specifying the communication pattern. Should be "paracrine" or "autocrine" if \code{ligand.type} is "lrpairs".}

\item{key.lrpairs}{A character string specifying the key ligand-receptor pair to filter the network. Only applicable if \code{ligand.type} is "lrpairs".}

\item{key.signal}{A character string specifying the key signaling molecule to filter the network. Required if \code{ligand.type} is "signal".}

\item{key.cluster}{A character string specifying specific clusters to include in the network. Only applicable when \code{comm.pattern} is "paracrine".}

\item{edge.size}{Numeric, controlling the thickness of edges based on communication probabilities.}

\item{edge.alpha}{Numeric, transparency level for edges, ranging from 0 to 1.}

\item{arrow.size}{Numeric, controlling the size of arrowheads on directed edges.}

\item{arrow.width}{Numeric, controlling the width of arrowheads on directed edges.}

\item{label.cex}{Numeric, scaling factor for vertex label font size.}

\item{label.distance}{Numeric, distance for vertex labels from the nodes in the chord diagram.}

\item{title.cex}{Numeric, scaling factor for title font size.}

\item{show.title}{Logical, indicating whether to display the plot title.}

\item{input.color}{A character vector of custom colors for nodes. Length should match the number of clusters. If NULL, default colors are used.}

\item{plot.type}{A character string specifying the type of plot to generate: "circle" for a circular plot or "chord" for a chord diagram.}
}
\value{
A network plot of the cell communication graph.
}
\description{
This function visualizes cell communication networks based on ligand-receptor pairs or signaling molecules.
It provides options to plot either paracrine or autocrine communication patterns in circular or chord plot types.
}
