% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingLRpairStats.R
\name{PlottingLRpairStats}
\alias{PlottingLRpairStats}
\title{Plot Ligand-Receptor Pair Statistics}
\usage{
PlottingLRpairStats(pcc_obj, thresh.p = 0.05, label.cex = 11, coord_flip = T)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{thresh.p}{Numeric, specifying the p-value threshold for significance. Default is 0.05.}

\item{label.cex}{Numeric, specifying the text size of axis labels in the plot.}

\item{coord_flip}{Logical, indicating whether to flip the x and y axes for better readability of long labels. Default is \code{TRUE}.}
}
\value{
A bar plot displaying the count of significant interactions for each ligand-receptor pair and signaling molecule type.
}
\description{
This function generates a bar plot of ligand-receptor pair statistics for different signaling molecules.
}
