% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingSignalContribution.R
\name{PlottingSignalContribution}
\alias{PlottingSignalContribution}
\title{Plot Signal Contribution bar plot}
\usage{
PlottingSignalContribution(
  pcc_obj,
  ligand.type = "lrpairs",
  key.signal = NULL,
  key.source = NULL,
  key.target = NULL,
  top.n = 10,
  label.cex = 10,
  title.cex = 11,
  expand.limits = -10,
  input.color = "#999999",
  show.title = T,
  coord.flip = F
)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{ligand.type}{A character string specifying the type of ligand, either "lrpairs" (ligand-receptor pairs) or "signal".}

\item{key.signal}{A character string specifying the key signaling molecule to filter the network.}

\item{key.source}{A character string specifying the cell type or tissue where the ligand located.}

\item{key.target}{A character string specifying the cell type or tissue where the receptor located.}

\item{label.cex}{Numeric, scaling factor for vertex label font size.}

\item{title.cex}{Numeric, scaling factor for title font size.}

\item{expand.limits}{Numeric, Expanding the plot limits on the y-axis. Default is -10.}

\item{input.color}{Character. Hex color code for the bars in the plot. Default is "#4682b4".}

\item{show.title}{Logical, indicating whether to display the plot title.}
}
\value{
A ggplot object representing the bar plot.
}
\description{
This function visualizes the contribution of ligand-receptor (LR) pairs or signal contributions.
}
