% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetPccIdent.R
\name{SetPccIdent}
\alias{SetPccIdent}
\title{Set the ident as the default in a plantcellchat object.}
\usage{
SetPccIdent(pcc_obj, input.ident)
}
\arguments{
\item{pcc_obj}{A PlantCellChat object.}

\item{input.ident}{The identifier to be set as the default label.}
}
\value{
A plantcellchat object after setting the default ident.
}
\description{
This function sets the specified \code{input.ident} as the default
label for the PlantCellChat object. It can be set either from the meta
information, if not found, as a cell label when the number of
unique cell labels matches the number of columns in the data matrix.
}
\examples{
\donttest{
pcc_obj <- SetPccIdent(pcc_obj, input.ident = "cell_type")
}
}
